+++ filament/src/Material.cpp
@@ -189,6 +189,7 @@ FMaterial::FMaterial(FEngine& engine, const Material::Builder& builder)
     parser->getRequiredAttributes(&mRequiredAttributes);
     parser->getRefractionMode(&mRefractionMode);
     parser->getRefractionType(&mRefractionType);
+    parser->getReflectionMode(&mReflectionMode);
 
     if (mBlendingMode == BlendingMode::MASKED) {
         parser->getMaskThreshold(&mMaskThreshold);
@@ -651,6 +652,10 @@ RefractionMode Material::getRefractionMode() const noexcept {
 
 RefractionType Material::getRefractionType() const noexcept {
     return upcast(this)->getRefractionType();
+}
+
+ReflectionMode Material::getReflectionMode() const noexcept {
+    return upcast(this)->getReflectionMode();
 }
 
 bool Material::hasParameter(const char* name) const noexcept {