+++ libs/filabridge/include/private/filament/UibStructs.h
@@ -136,8 +136,16 @@ struct PerViewUib { // NOLINT(cppcoreguidelines-pro-type-member-init)
     float nearOverFarMinusNear;         // n / (f-n), always positive
     float temporalNoise;                // noise [0,1] when TAA is used, 0 otherwise
 
+    // Screen-space reflections
+    math::mat4f ssrReprojection;
+    math::mat4f ssrProjectToPixelMatrix;
+    float ssrThickness;                 // ssr thickness, in world units
+    float ssrBias;                      // ssr bias, in world units
+    float ssrDistance;                  // ssr world raycast distance, 0 when ssr is off
+    float ssrStride;                    // ssr texel stride, >= 1.0
+
     // bring PerViewUib to 2 KiB
-    math::float4 padding2[58];
+    math::float4 padding3[49];
 };
 
 // 2 KiB == 128 float4s