+++ libs/filamat/src/UibGenerator.cpp
@@ -117,8 +117,16 @@ UniformInterfaceBlock const& UibGenerator::getPerViewUib() noexcept  {
             .add("nearOverFarMinusNear",    1, UniformInterfaceBlock::Type::FLOAT, Precision::HIGH)
             .add("temporalNoise",           1, UniformInterfaceBlock::Type::FLOAT, Precision::HIGH)
 
+            // Screen-space reflection parameters
+            .add("ssrReprojection",         1, UniformInterfaceBlock::Type::MAT4, Precision::HIGH)
+            .add("ssrProjectToPixelMatrix", 1, UniformInterfaceBlock::Type::MAT4, Precision::HIGH)
+            .add("ssrThickness",            1, UniformInterfaceBlock::Type::FLOAT)
+            .add("ssrBias",                 1, UniformInterfaceBlock::Type::FLOAT)
+            .add("ssrDistance",             1, UniformInterfaceBlock::Type::FLOAT)
+            .add("ssrStride",               1, UniformInterfaceBlock::Type::FLOAT)
+
             // bring PerViewUib to 2 KiB
-            .add("padding2", 58, UniformInterfaceBlock::Type::FLOAT4)
+            .add("padding3", 49, UniformInterfaceBlock::Type::FLOAT4)
             .build();
     return uib;
 }